Add data source map name to host model
The data source map is the name for the device used in the ganglia
output. This is an external name and not always suitable for customer
consumption.

Currently, it is the same as the host's device name, later work will
provide a lookup for the data source map from the device name.