Lookup map to host from device name
When adding the first metric for a device lookup the device's data
source map to host.  This allows us to expose the already exposed device
name in the API while using the correct map to host.

Currently, the map to host is looked up from a JSON file.  The JSON
file needs to be created and updated by an external script.