Change encoding in GDS output
It appears that if gmetad receives documents with differing encodings it
drops some of the documents.  As our other GDS daemons use ISO-8859-1
mrapi is updated to do so too.

Also, the doctype has been added.  I don't think that this is necessary
but was done during discovery of the encoding issue above.  I can't see
any reason to remove this now that it has been added.