API to create authentication tokens
The goal here is for Emma to have access to the shared secret, to
create its own JWT token which it passes with the request to create
tokens and for mrd to authenticate the request to create tokens.

On further consideration, perhaps we should just have emma create all of
the tokens.