Record results via a Ruby script
The existing apps expect the results to be stored in memcache using Ruby
Marhsal for the serialisation.  Golang has poor support for that
serialisation format, so instead we use a Ruby script to serialise and
store the results.

The ruby script is given the results to store as JSON via stdin.