Add ticker struct permitting manual ticking
Ticker is a wrapper around a time.Ticker.  In addition to the periodic
ticks provided by time.Ticker, it also has a TickNow() method which
causes an immediate attempt to tick.  If that tick is successful, the
wrapped time.Ticker is reset with its existing duration.  The reasons
that a call to TickNow might not cause a successful tick are documented
with that method.