Metrics are reported via device id not device name
Using the device id allows reporting metrics for devices with duplicate
names.  As the metrics are still processed by ganglia, there is a need
for each host to have a unique name (or at least unique to its cluster
and we only use a single cluster). The solution is to make the metrics
available to ganglia using the device's memcache key, e.g.,
`hacor:device:<ID>`.

So the process is

1. HTTP API accepts metric reported via device ID.
2. Device ID is used to lookup the data source map, e.g.,
   `unspecified/unspecified/hacor:device:3`.
3. Metrics are reported to ganglia with that data source map.
4. Metrics are retrieved from ganglia for further processing and
   reported to memcache using the device's memcache key, e.g.,
   `hacor:device:3`.

This has required some changes in the memcache structure created by
ct-visualisation-app.