`/metrics/historic` returns list of historic metrics
The only data that we have available is the metric name.  We have lost
the nature and units and aggregations such as minimum and maximum make
little sense and would probably be too expensive to calculate.

Whilst we could probably get the nature and units from the current
metrics so could the client.