Don't run ticker tests in docker container
They can be somewhat flaky due to testing times deltas.  The flakiness
could be due to how I'm developing the new ansible playbook to build all
of the concertim components.  But for now they are going to be disabled
in the docker build.

Speed up docker builds; add cache bust arg.