Create and update our own RRD files
Before we can remove ganglia we need to be able to create and update or
own RRD files.  This commit does that.  RRD files are still created by
ganglia, but they are now ignored.

The directory hierarchy for our RRD files is the same as that used by
ganglia.

The RRD step and RRAs can be specified in the config file.