Fix issue with LastProcessed
Modifications to the pending metrics whilst processing them are lost
immediately following the completion of the processing run.  To avoid
losing the changes to LastProcessed, the pending repository interface
has gained an UpdateLastProcessed(HostId, MetricName, time.Time) method.

As this required a change to the pending repository implementation, I
took the opportunity to move it out of `repositories/memory` and into
`inmem`, and also to simplfy its implementation.